Friedman benchmark data

Simulate data from the Friedman 1 benchmark problem. See mlbench.friedman1 for details and references.

gen_friedman(
  n_samples = 100,
  n_features = 10,
  n_bins = NULL,
  sigma = 0.1,
  seed = NULL
)

Arguments

n_samples
Integer specifying the number of samples (i.e., rows) to generate. Default is 100.

n_features
Integer specifying the number of features to generate. Default is 10.

n_bins
Integer specifying the number of (roughly) equal sized bins to split the response into. Default is NULL for no binning. Setting to a positive integer > 1 effectively turns this into a classification problem where n_bins gives the number of classes.

sigma
Numeric specifying the standard deviation of the noise.

seed
Integer specifying the random seed. If NULL (the default) the results will be different each time the function is run.

Note

This function is mostly used for internal testing.

Examples

gen_friedman()

#>             y          x1         x2         x3          x4          x5
#> 1    8.655504 0.221668986 0.03635446 0.13086855 0.344219954 0.484186048
#> 2   21.932600 0.725366318 0.57582960 0.09396259 0.400690940 0.980629443
#> 3   15.431864 0.490189948 0.50251175 0.42502920 0.561758569 0.579596978
#> 4   23.923772 0.835470283 0.60702179 0.19535789 0.748671596 0.911223026
#> 5   20.605801 0.991906658 0.22286123 0.80980792 0.906811133 0.625211999
#> 6   18.554281 0.346081430 0.91616168 0.85473985 0.271615403 0.979046850
#> 7   17.645488 0.418975663 0.82791546 0.28983882 0.372759387 0.837340214
#> 8   18.909681 0.322696026 0.68803933 0.67133869 0.742115303 0.873637285
#> 9   17.372453 0.396875542 0.35150256 0.10018998 0.813145185 0.383432061
#> 10  13.636122 0.488679143 0.45159015 0.16553686 0.104501814 0.792597824
#> 11  19.462026 0.614113803 0.72719079 0.79259980 0.541026698 0.494623203
#> 12   8.022182 0.063667870 0.70307441 0.03750283 0.194649176 0.100877792
#> 13  11.470856 0.074066723 0.17103532 0.20133683 0.927386577 0.049154697
#> 14  16.302362 0.155926090 0.84952658 0.15135914 0.638401545 0.695484582
#> 15  21.605096 0.923669656 0.65001947 0.24267098 0.675596890 0.765460214
#> 16   3.242614 0.004661539 0.99127801 0.20511504 0.117083271 0.020538517
#> 17  16.389842 0.936478112 0.41511526 0.25662458 0.397384518 0.371077226
#> 18  12.876670 0.011452872 0.14028739 0.83268511 0.861546160 0.407181462
#> 19  10.862174 0.470285797 0.23368954 0.30251086 0.510669256 0.319029736
#> 20  17.643459 0.798402437 0.62085847 0.16461916 0.178023820 0.736263002
#> 21  13.437165 0.379698449 0.59262424 0.68564825 0.587881071 0.069607003
#> 22  14.799559 0.344868697 0.57378609 0.53792722 0.759753137 0.244497828
#> 23  15.249655 0.451476743 0.84584926 0.56389268 0.361598653 0.447154938
#> 24  17.237459 0.727334189 0.39070353 0.92600958 0.536382562 0.058174848
#> 25  20.556724 0.603029140 0.57258133 0.86343377 0.830891874 0.162436850
#> 26  14.830297 0.560185106 0.55231124 0.29884191 0.349775972 0.442897327
#> 27  14.881873 0.681157269 0.97309901 0.11267874 0.096285560 0.449267249
#> 28  16.651718 0.387125138 0.28380914 0.99488427 0.670913033 0.376186142
#> 29  16.305505 0.418990348 0.69877837 0.36798670 0.456576811 0.688399482
#> 30   8.231916 0.973969692 0.18928457 0.37347093 0.128433277 0.274051381
#> 31   9.750062 0.190144064 0.08930799 0.66835274 0.476358568 0.784298365
#> 32  13.465712 0.067567294 0.57649553 0.13285778 0.560884351 0.770395850
#> 33  12.131595 0.319376916 0.13998135 0.62095492 0.648698271 0.763465408
#> 34   9.704876 0.259054112 0.84163036 0.33750117 0.193796872 0.203534953
#> 35   7.805929 0.323437379 0.15535833 0.73558925 0.086424554 0.852507567
#> 36  18.916844 0.899707715 0.26853609 0.37227316 0.832128844 0.696536337
#> 37  17.594090 0.112210827 0.30568118 0.85752452 0.936493818 0.912019301
#> 38  11.301212 0.409680396 0.46660604 0.25914505 0.002675714 0.878405672
#> 39  17.060164 0.785769888 0.46944648 0.01864676 0.091631239 0.472508086
#> 40   9.739774 0.242360523 0.57468389 0.47782845 0.499938722 0.090770660
#> 41  12.652981 0.415976827 0.24623651 0.96326295 0.490060022 0.042620456
#> 42  10.769409 0.503456446 0.08997576 0.42849312 0.838377257 0.156039725
#> 43  14.685778 0.379893373 0.31266987 0.85368988 0.639341023 0.426826963
#> 44  24.037147 0.536513995 0.88267746 0.21480870 0.964476610 0.586039324
#> 45  15.787838 0.725832683 0.18644344 0.78838182 0.996954729 0.002818695
#> 46  19.380804 0.980947863 0.46615485 0.86390519 0.554366705 0.237893737
#> 47  13.483567 0.102070478 0.92087224 0.05022977 0.459220991 0.364775082
#> 48  11.334212 0.098966519 0.72566312 0.53063293 0.705915459 0.424781820
#> 49  25.074363 0.765224977 0.77184499 0.20080313 0.876003243 0.992260950
#> 50  17.448904 0.208629725 0.07981846 0.98664681 0.881520300 0.669954233
#> 51  13.268990 0.501132404 0.41042764 0.21114206 0.431142829 0.264953824
#> 52  18.040076 0.583593405 0.39001950 0.32992431 0.968491144 0.238466105
#> 53  17.432746 0.881847649 0.53256331 0.71140765 0.601880318 0.120980049
#> 54  17.868863 0.536537957 0.98933363 0.37950479 0.523988937 0.460440699
#> 55  20.844462 0.993123762 0.32153434 0.80654159 0.899823000 0.298816899
#> 56  19.801157 0.986969389 0.29429222 0.15296603 0.825275206 0.256206517
#> 57  11.964502 0.119513957 0.29407653 0.79473830 0.627658969 0.541484194
#> 58  17.262163 0.724024503 0.28901263 0.76244150 0.677875768 0.620547654
#> 59   9.597301 0.076432820 0.74939434 0.87249380 0.488453067 0.010839765
#> 60  12.850182 0.273731682 0.45072427 0.82981844 0.596002576 0.184051059
#> 61  17.549209 0.563140662 0.68482816 0.83521187 0.157229910 0.890548117
#> 62  20.954097 0.633241213 0.47039534 0.63585414 0.781083628 0.926599846
#> 63   7.675162 0.111568248 0.92894803 0.86182972 0.134930611 0.121385235
#> 64   7.222767 0.102505034 0.12024477 0.61082614 0.307622932 0.718637321
#> 65  16.881140 0.889974394 0.87320225 0.38555337 0.979586648 0.062622033
#> 66  17.194403 0.939933171 0.06033869 0.93589841 0.654561191 0.995727064
#> 67  14.003084 0.688927307 0.21496477 0.61650054 0.829796989 0.188294070
#> 68  13.690262 0.352700548 0.83979186 0.08811323 0.202330501 0.021569320
#> 69  10.712303 0.030351268 0.63144225 0.42837104 0.887481175 0.212353996
#> 70  10.514974 0.062016443 0.60136958 0.53919921 0.459258048 0.918380492
#> 71   5.735426 0.480150337 0.11125568 0.48867419 0.239120423 0.321068264
#> 72   7.783056 0.057453917 0.58320457 0.92162196 0.111072880 0.394305044
#> 73  11.593534 0.250723377 0.06263289 0.44575737 0.987451725 0.259606398
#> 74   8.699722 0.252691907 0.19835347 0.38034625 0.632278231 0.083224106
#> 75   8.150382 0.178551870 0.99026670 0.40649684 0.033040064 0.482496028
#> 76   8.709118 0.024010095 0.82844425 0.56140900 0.436779671 0.718148117
#> 77  16.456329 0.304309153 0.45977917 0.57231572 0.986824797 0.461918383
#> 78   8.814603 0.294217970 0.20654807 0.19824140 0.105087175 0.837478388
#> 79  15.074774 0.130726798 0.90934051 0.32149008 0.587852684 0.981739305
#> 80  18.420242 0.602175667 0.31997205 0.78413755 0.753858332 0.705660478
#> 81  13.961009 0.611169651 0.88241024 0.32110931 0.250050217 0.129770564
#> 82  11.752398 0.173902044 0.47459594 0.07139317 0.330094106 0.421915213
#> 83   5.628336 0.939820190 0.03725514 0.76589037 0.279013717 0.079877399
#> 84  20.854360 0.985659893 0.66010465 0.84230280 0.548309589 0.852157482
#> 85  18.243630 0.710645568 0.26293216 0.37449341 0.881926132 0.774795359
#> 86  15.083364 0.773631262 0.72091023 0.68814952 0.242998869 0.442551188
#> 87  13.300913 0.747659922 0.11256950 0.62695584 0.665718448 0.744111299
#> 88  16.394557 0.228787420 0.63855662 0.97265856 0.726107717 0.043384629
#> 89  17.917545 0.417181292 0.94818527 0.50573190 0.614964551 0.481667765
#> 90  13.420221 0.208818242 0.36715935 0.99843660 0.118780760 0.968757692
#> 91  17.087078 0.240902313 0.91396841 0.99946119 0.156040334 0.814385115
#> 92  12.369700 0.778687984 0.15189164 0.03325135 0.103686596 0.677300980
#> 93  11.896147 0.925297623 0.27046428 0.32785408 0.333307283 0.153235649
#> 94  14.311125 0.177202225 0.58295494 0.01403061 0.170534124 0.951172760
#> 95  21.627191 0.521641018 0.54877747 0.19344360 0.972996171 0.424844675
#> 96   4.900746 0.130689902 0.04630522 0.19053992 0.257813476 0.034775290
#> 97  11.589380 0.039007949 0.04051201 0.72013613 0.812262412 0.508475263
#> 98  22.314187 0.846004942 0.43269853 0.43426729 0.867670665 0.890939787
#> 99  16.394765 0.713334864 0.92431874 0.79783747 0.310005208 0.553998113
#> 100 20.840748 0.877146576 0.85922517 0.93531709 0.701122221 0.615019433
#>              x6         x7          x8         x9         x10
#> 1   0.857460270 0.79283686 0.868148895 0.12211754 0.631894944
#> 2   0.400757526 0.61667009 0.914441498 0.25927137 0.040752097
#> 3   0.072330632 0.68356623 0.553648580 0.43021381 0.488671324
#> 4   0.163148977 0.19917807 0.869617630 0.96539996 0.095381634
#> 5   0.491292784 0.16856193 0.299087636 0.96826067 0.653653607
#> 6   0.840894020 0.40088800 0.491647091 0.82735333 0.925639774
#> 7   0.535876179 0.78679475 0.655921439 0.20756668 0.239565232
#> 8   0.369983375 0.08740342 0.880514636 0.58394584 0.583477433
#> 9   0.683882175 0.04850435 0.373707469 0.20597122 0.066567236
#> 10  0.669345597 0.83878296 0.870118046 0.78291292 0.424344913
#> 11  0.505187855 0.83415471 0.735768442 0.68350222 0.138601692
#> 12  0.011409613 0.80274473 0.529878760 0.89138062 0.309986631
#> 13  0.791778329 0.80007409 0.561719570 0.31755195 0.329966714
#> 14  0.933460541 0.65943775 0.101710201 0.51663575 0.657190276
#> 15  0.580039089 0.21314219 0.752276375 0.86525574 0.486444642
#> 16  0.491637964 0.34346406 0.376827343 0.36553513 0.010334489
#> 17  0.800024862 0.27696087 0.876567017 0.01371411 0.574739437
#> 18  0.349688162 0.99784685 0.888411286 0.92394768 0.774835635
#> 19  0.558092570 0.87376923 0.578186824 0.67538078 0.092092840
#> 20  0.380757889 0.31451237 0.492116883 0.32250059 0.156985706
#> 21  0.006039485 0.83228730 0.458276265 0.73555276 0.485941888
#> 22  0.856134011 0.43697321 0.421402605 0.23500804 0.472186121
#> 23  0.951483841 0.22315713 0.200728744 0.68826858 0.078361105
#> 24  0.675308276 0.29388392 0.202511362 0.27184364 0.187036444
#> 25  0.974488168 0.41380485 0.600831289 0.90840122 0.491624003
#> 26  0.935298964 0.87020097 0.732319857 0.67706587 0.368052977
#> 27  0.320311913 0.38389495 0.229237816 0.02368357 0.965941215
#> 28  0.270332079 0.73133957 0.714826101 0.49632615 0.042077209
#> 29  0.843963213 0.36823945 0.252073761 0.08034507 0.087678315
#> 30  0.287053454 0.17589042 0.622353352 0.56297710 0.404373041
#> 31  0.766257521 0.02361789 0.743724781 0.71135307 0.551258451
#> 32  0.686546477 0.38230111 0.267249618 0.30285779 0.215913408
#> 33  0.216392208 0.35383314 0.114639532 0.50344148 0.712149488
#> 34  0.700999162 0.97115229 0.901311499 0.05498463 0.811910261
#> 35  0.624936773 0.70611076 0.842161050 0.32129493 0.465495184
#> 36  0.047174204 0.98137062 0.865132697 0.65672054 0.440406783
#> 37  0.668185351 0.16470426 0.451973337 0.71181576 0.697172235
#> 38  0.946628477 0.84712852 0.663471942 0.37712442 0.827695406
#> 39  0.112816765 0.73879712 0.727030973 0.27375231 0.817774341
#> 40  0.701326447 0.72700361 0.567646981 0.91458159 0.948268371
#> 41  0.945768989 0.10500108 0.627580963 0.18762806 0.557828719
#> 42  0.159300349 0.32578482 0.989619033 0.02873024 0.636543262
#> 43  0.052139587 0.37879508 0.704422678 0.67015157 0.766766402
#> 44  0.147726096 0.15512109 0.352750666 0.03936991 0.605559197
#> 45  0.697298142 0.66307524 0.216439642 0.60776262 0.711862320
#> 46  0.057532220 0.94076881 0.391868545 0.24320549 0.269936275
#> 47  0.821641564 0.86082793 0.297987778 0.54900831 0.230596611
#> 48  0.669226542 0.58801192 0.441946210 0.39492145 0.076440749
#> 49  0.560662086 0.65521014 0.889416563 0.11676066 0.210054127
#> 50  0.069965977 0.54429532 0.943514114 0.90802893 0.010796506
#> 51  0.698230357 0.56817940 0.331379056 0.69789402 0.402369910
#> 52  0.448798469 0.52736736 0.076415496 0.15053377 0.100735299
#> 53  0.181841866 0.60808073 0.469338949 0.67299070 0.646176061
#> 54  0.016198873 0.48163633 0.435353419 0.09926794 0.275696628
#> 55  0.646149317 0.83476121 0.577558230 0.15656788 0.250505514
#> 56  0.830915296 0.49740223 0.737186983 0.79028514 0.196223264
#> 57  0.561636838 0.07739889 0.791786693 0.26290476 0.249278289
#> 58  0.853955714 0.67809567 0.709059633 0.31900153 0.055292644
#> 59  0.883929934 0.72281911 0.002721401 0.17883548 0.978355116
#> 60  0.287696199 0.07441112 0.836529555 0.66701564 0.519116459
#> 61  0.465752248 0.46823574 0.442997099 0.36808396 0.568000829
#> 62  0.968346498 0.16173538 0.445749999 0.11839093 0.404718738
#> 63  0.353765457 0.90754789 0.199521945 0.22950933 0.415659736
#> 64  0.320621889 0.60144087 0.271259413 0.65387585 0.323417492
#> 65  0.747200130 0.53311287 0.146254664 0.25595786 0.834168654
#> 66  0.651896443 0.98525061 0.486490316 0.37177659 0.104867197
#> 67  0.122993345 0.72234537 0.077065458 0.94549853 0.930466204
#> 68  0.107624776 0.97841470 0.954752345 0.94839389 0.088616451
#> 69  0.000884854 0.17220105 0.691548009 0.22663052 0.069306283
#> 70  0.881172108 0.08979754 0.284009779 0.75313725 0.401476627
#> 71  0.489369652 0.74689687 0.802858111 0.64782546 0.209911534
#> 72  0.265625506 0.85154102 0.245812844 0.11602550 0.100887767
#> 73  0.213899547 0.79989985 0.238066752 0.40853985 0.702334291
#> 74  0.063495589 0.13748396 0.629413611 0.81284641 0.292434771
#> 75  0.575380990 0.13054212 0.313641107 0.75117409 0.701813064
#> 76  0.172447122 0.75329023 0.126355306 0.18380117 0.650995285
#> 77  0.230374837 0.35357929 0.767324975 0.25544895 0.790341998
#> 78  0.247475870 0.08624698 0.500635321 0.42410824 0.224970208
#> 79  0.201049262 0.50149100 0.175946329 0.97181170 0.502161956
#> 80  0.554377512 0.84180558 0.435213814 0.38485597 0.047613695
#> 81  0.218384935 0.40057627 0.653111830 0.62917164 0.296823329
#> 82  0.544565327 0.38936845 0.628993650 0.58188505 0.014406463
#> 83  0.523222651 0.19547194 0.885172290 0.66844840 0.248675680
#> 84  0.565244167 0.76477694 0.109743171 0.68730054 0.231467532
#> 85  0.291245153 0.99734160 0.858654720 0.11622300 0.846493064
#> 86  0.992371926 0.31189263 0.666766479 0.22545228 0.336746750
#> 87  0.366282751 0.01303621 0.258605690 0.62842420 0.005534482
#> 88  0.384523053 0.81684094 0.757495314 0.59406343 0.434733107
#> 89  0.335552482 0.52092851 0.179436560 0.58364528 0.804479253
#> 90  0.708112209 0.39749206 0.083145949 0.36882164 0.193645956
#> 91  0.349974752 0.03720775 0.224183933 0.79623287 0.367928172
#> 92  0.508430470 0.25650660 0.522984717 0.80578815 0.536768902
#> 93  0.328598775 0.32772904 0.280462991 0.21018280 0.623037692
#> 94  0.101532507 0.25012448 0.867713927 0.54118150 0.635705414
#> 95  0.051488115 0.77860328 0.360458912 0.23634708 0.166825783
#> 96  0.106177686 0.77931811 0.823581273 0.44153819 0.869693672
#> 97  0.884073677 0.77888356 0.963321767 0.29522256 0.848551975
#> 98  0.825238326 0.24364548 0.161241072 0.60435545 0.502786030
#> 99  0.289900429 0.21980740 0.367676386 0.83077099 0.161630602
#> 100 0.613827475 0.12738871 0.406286990 0.26632076 0.301377810